refactor(colocation): move app layout and post giscus
