+++ src/app/layout.tsx
@@ -1,11 +1,11 @@
 import "@semantic/styles/globals.css";
 
-import { Layout } from "@semantic/components/layout/root";
 import { METADATA } from "@semantic/constants/metadata";
 import type { Metadata, Viewport } from "next";
 import { ThemeProvider } from "next-themes";
 import type { PropsWithChildren } from "react";
 import { twMerge } from "tailwind-merge";
+import { Layout } from "./_components/layout/root";
 
 import { GeistMono, Pretendard } from "./_fonts";
